chore: code refactory
