+++ key-vault/src/db/errors.rs
