+++ key-vault/src/utilities/tests.rs
@@ -23,7 +23,7 @@ fn test_fail_encrypt_decrypt() {
 
 #[test]
 fn test_zeroize_on_drop_decrypt_output() {
-  use super::secure_vec::ZEROIZED;
+  use crate::secure_vec::ZEROIZED;
   ZEROIZED.store(false, Ordering::SeqCst);
   let password = vec![1, 2, 3];
   let data = b"test";
